Moved VirtueRS3-specific classes out of the cache package
