+++ src/main/java/org/virtue/config/quickchatphrasetype/QuickChatPhraseType.java
@@ -19,7 +19,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package org.virtue.cache.def.impl;
+package org.virtue.config.quickchatphrasetype;
 
 import java.io.IOException;
 import java.nio.ByteBuffer;
